move generateFailedExecutionFromError to ExecutionDataService
