+++ packages/cli/src/executions/execution.service.ts
@@ -10,7 +10,6 @@ import type {
 	IWorkflowBase,
 	WorkflowExecuteMode,
 	IWorkflowExecutionDataProcess,
-	IRun,
 } from 'n8n-workflow';
 import {
 	ExecutionStatusList,
@@ -524,62 +523,5 @@ export class ExecutionService {
 		if (updateData.tags) {
 			await this.annotationTagMappingRepository.overwriteTags(annotation.id, updateData.tags);
 		}
-	}
-
-	generateFailedExecutionFromError(
-		mode: WorkflowExecuteMode,
-		error: ExecutionError,
-		node: INode | undefined,
-		startTime = Date.now(),
-	): IRun {
-		const executionError = {
-			...error,
-			message: error.message,
-			stack: error.stack,
-		};
-		const returnData: IRun = {
-			data: {
-				resultData: {
-					error: executionError,
-					runData: {},
-				},
-			},
-			finished: false,
-			mode,
-			startedAt: new Date(),
-			stoppedAt: new Date(),
-			status: 'error',
-		};
-
-		if (node) {
-			returnData.data.startData = {
-				destinationNode: node.name,
-				runNodeFilter: [node.name],
-			};
-			returnData.data.resultData.lastNodeExecuted = node.name;
-			returnData.data.resultData.runData[node.name] = [
-				{
-					startTime,
-					executionTime: 0,
-					executionStatus: 'error',
-					error: executionError,
-					source: [],
-				},
-			];
-			returnData.data.executionData = {
-				contextData: {},
-				metadata: {},
-				waitingExecution: {},
-				waitingExecutionSource: {},
-				nodeExecutionStack: [
-					{
-						node,
-						data: {},
-						source: null,
-					},
-				],
-			};
-		}
-		return returnData;
 	}
 }
