+++ packages/cli/src/workflow-execute-additional-data.ts
@@ -37,7 +37,7 @@ import { WorkflowRepository } from '@/databases/repositories/workflow.repository
 import { EventService } from '@/events/event.service';
 import type { AiEventMap, AiEventPayload } from '@/events/maps/ai.event-map';
 import { getLifecycleHooksForSubExecutions } from '@/execution-lifecycle/execution-lifecycle-hooks';
-import { ExecutionService } from '@/executions/execution.service';
+import { ExecutionDataService } from '@/executions/execution-data.service';
 import {
 	CredentialsPermissionChecker,
 	SubworkflowPolicyChecker,
@@ -262,7 +262,7 @@ async function startExecution(
 		data = await execution;
 	} catch (error) {
 		const executionError = error as ExecutionError;
-		const fullRunData = Container.get(ExecutionService).generateFailedExecutionFromError(
+		const fullRunData = Container.get(ExecutionDataService).generateFailedExecutionFromError(
 			runData.executionMode,
 			executionError,
 			'node' in executionError ? executionError.node : undefined,
