+++ packages/cli/src/workflow-runner.ts
@@ -27,7 +27,7 @@ import {
 	getLifecycleHooksForScalingWorker,
 	getLifecycleHooksForScalingMain,
 } from '@/execution-lifecycle/execution-lifecycle-hooks';
-import { ExecutionService } from '@/executions/execution.service';
+import { ExecutionDataService } from '@/executions/execution-data.service';
 import { CredentialsPermissionChecker } from '@/executions/pre-execution-checks';
 import { ManualExecutionService } from '@/manual-execution.service';
 import { NodeTypes } from '@/node-types';
@@ -52,7 +52,7 @@ export class WorkflowRunner {
 		private readonly credentialsPermissionChecker: CredentialsPermissionChecker,
 		private readonly instanceSettings: InstanceSettings,
 		private readonly manualExecutionService: ManualExecutionService,
-		private readonly executionService: ExecutionService,
+		private readonly executionDataService: ExecutionDataService,
 	) {}
 
 	/** The process did error */
@@ -137,7 +137,7 @@ export class WorkflowRunner {
 			await this.credentialsPermissionChecker.check(workflowId, nodes);
 		} catch (error) {
 			// Create a failed execution with the data for the node, save it and abort execution
-			const runData = this.executionService.generateFailedExecutionFromError(
+			const runData = this.executionDataService.generateFailedExecutionFromError(
 				data.executionMode,
 				error,
 				error.node,
